branches/main: replaced some int/unsigned int by stride_type/size_type in iterator classes
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@1571 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/include/cisstVector/vctDynamicVector.h
@@ -65,7 +65,7 @@ http://www.cisst.org/cisst/license.txt.
 
   // the vectorSize variable can be set to any value at any time
   // before creating the vector.
-  unsigned int vectorSize = 12;
+  size_t vectorSize = 12;
 
   // constructor allocation
   vctDynamicVector<ElementType> v1(vectorSize);
